requirements.txt
decord>=0.6.0
fvcore
matplotlib
mmcv-full==1.3.18
mmdet==2.23.0
mmpose==0.24.0
moviepy
numpy>=1.19.5
opencv-contrib-python
opencv-python
pymemcache
scipy
torch>=1.5
tqdm
